refactor: 도메인으로 분리
